Moved all Input Data classes to use_case
