+++ src/main/java/use_case/user/SignUpInputBoundary.java
@@ -1,7 +1,5 @@
 package use_case.user;
 
-import input_data.SignUpInputData;
-
 public interface SignUpInputBoundary {
     void createUser(SignUpInputData inputData);
     void switchtoLogin();
